minimap creation start
I've followed the HowTo but don't completely succeed. The minimap appears but seems not well positionned. I think that the issue comes from the parameters in the DLC mounter of the reddlc file.
Moreover the world map is still full black when we want to see it via the menu.